requirements.txt
altgraph==0.17.4
ansicon==1.89.0
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
auto-py-to-exe==2.43.3
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
console==0.9907
contourpy==1.2.0
cryptography==42.0.8
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.8
Django==3.2.5
django-browser-reload==1.12.1
django-cors-headers==4.3.0
django-filter==23.5
django-livesync==0.5
django-multiselectfield==0.1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Eel==0.16.0
ezenv==0.92
Flask==3.0.3
fonttools==4.49.0
fpdf==1.7.2
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
html5lib==1.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
jinxed==1.2.1
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.3
matplotlib==3.8.3
numpy==1.24.3
oscrypto==1.3.0
packaging==23.2
pandas==2.0.1
pdfkit==1.0.0
pdfrw==0.4
peewee==3.17.6
pefile==2023.2.7
Pillow==9.4.0
proxy-tools==0.1.0
psycopg2==2.9.5
pyarmor==8.4.7
pyarmor.cli.core==5.4.3
pycparser==2.21
pydyf==0.11.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.6.0
pyparsing==3.1.2
pypdf==4.3.0
pyphen==0.15.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySide6_Essentials==6.7.2
python-bidi==0.4.2
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2024.1
pywebview==4.4.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.31.0
shiboken6==6.7.2
shutup==0.2.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.3.0
tornado==6.4
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.3
urllib3==2.0.7
watchdog==3.0.0
weasyprint==62.3
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
whichcraft==0.6.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.16
zope.event==5.0
zope.interface==6.2
zopfli==0.2.3
